# Conflicts:
#	Makefile
#	lib/libft/Makefile
#	src/main.c

+++ lib/libft/src/io/ft_putptr.c
@@ -14,5 +14,5 @@
 
 int	ft_putptr(void *ptr)
 {
-	return (ft_putstr("0x") + ft_putlhex((unsigned long long)ptr));
+	return (ft_putstr("0x") + ft_putlhex((unsigned long)ptr));
 }
